refactor: archive 도메인 생성
