+++ bookduck/src/main/java/com/mmc/bookduck/domain/archive/dto/request/ExcerptCreateRequestDto.java
@@ -1,8 +1,8 @@
-package com.mmc.bookduck.domain.excerpt.dto.request;
+package com.mmc.bookduck.domain.archive.dto.request;
 
 import com.mmc.bookduck.domain.book.entity.UserBook;
 import com.mmc.bookduck.domain.common.Visibility;
-import com.mmc.bookduck.domain.excerpt.entity.Excerpt;
+import com.mmc.bookduck.domain.archive.entity.Excerpt;
 import com.mmc.bookduck.domain.user.entity.User;
 
 public record ExcerptCreateRequestDto(
