+++ bookduck/src/main/java/com/mmc/bookduck/domain/archive/repository/ExcerptRepository.java
@@ -1,6 +1,6 @@
-package com.mmc.bookduck.domain.excerpt.repository;
+package com.mmc.bookduck.domain.archive.repository;
 
-import com.mmc.bookduck.domain.excerpt.entity.Excerpt;
+import com.mmc.bookduck.domain.archive.entity.Excerpt;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ExcerptRepository extends JpaRepository<Excerpt, Long> {
